fix(generic): update cloudflare challenge detection
- added bypassPage (NeloManga returns bad data when CF header is present and referer = /)
